feat(map): add interactive rotation and auto-centering for selected country
