+++ .changeset/every-trains-divide.md
@@ -3,3 +3,4 @@
 ---
 
 feat(worldmap): add globe rotation and interactive dragging logic
+Add interactive drag rotation and auto-centering to the 3D globe, powered by a new `useGlobeRotation` custom hook.
